+++ src/components/spatial/utils.js
@@ -1,0 +1,34 @@
+import shortNumber from 'short-number';
+import { pluralize } from '../utils';
+
+/**
+ * Make a subtitle for the spatial component.
+ * @param {object} params
+ * @param {number} params.observationsCount
+ * @param {string} params.observationsLabel
+ * @param {string} params.observationsPluralLabel
+ * @param {number} params.subobservationsCount
+ * @param {string} params.subobservationsLabel
+ * @param {string} params.subobservationsPluralLabel
+ * @param {number} params.locationsCount
+ * @returns {string} The subtitle string,
+ * with info about items with zero counts omitted.
+ */
+export function makeSpatialSubtitle({
+  observationsCount, observationsLabel, observationsPluralLabel,
+  subobservationsCount, subobservationsLabel, subobservationsPluralLabel,
+  locationsCount,
+}) {
+  const parts = [];
+  if (observationsCount > 0) {
+    parts.push(`${observationsCount} ${pluralize(observationsLabel, observationsPluralLabel, observationsCount)}`);
+  }
+  if (subobservationsCount > 0) {
+    let part = `${subobservationsCount} ${pluralize(subobservationsLabel, subobservationsPluralLabel, subobservationsCount)}`;
+    if (locationsCount > 0) {
+      part += ` at ${shortNumber(locationsCount)} locations`;
+    }
+    parts.push(part);
+  }
+  return parts.join(', ');
+}